One hour to interview

+++ src/main/java/com/jafar/wiring/controllers/Controller.java
@@ -15,49 +15,47 @@ import java.util.ArrayList;
 import java.util.Arrays;
 
 @RestController
+@CrossOrigin("*")
 public class Controller {
 
     @Value("${file.upload-dir}")
     private String FILE_DIRECTORY;
 
-    private String UPLOADED_FILE_DIRECTORY;
 
-    @PostMapping("/upload-dataset")
-    private String uploadDataset(@RequestParam("File") MultipartFile paramFile) {
+    @PostMapping("/process-dataset")
+    private DatasetEndpoint uploadDataset(@RequestParam("file") MultipartFile paramFile, @RequestParam("flag") boolean flag) {
         try {
+            System.out.println("Processing file " + paramFile.getOriginalFilename());
             File file = new File(FILE_DIRECTORY + paramFile.getOriginalFilename());
-            if(!file.createNewFile())
-                return "Error: File could not be created!!!";
-            FileOutputStream fos = new FileOutputStream(file);
-            fos.write(paramFile.getBytes());
-            fos.close();
-            UPLOADED_FILE_DIRECTORY = FILE_DIRECTORY + paramFile.getOriginalFilename();
-            return "File uploaded successfully. :D";
+            boolean fileExists = file.exists();
+
+            if(!fileExists) {
+                if(!file.createNewFile()) {
+                    System.out.println("Something went wrong");
+                    return null;
+                }
+                FileOutputStream fos = new FileOutputStream(file);
+                fos.write(paramFile.getBytes());
+                fos.close();
+                String uploadedFileDirectory = FILE_DIRECTORY + paramFile.getOriginalFilename();
+                Processor processor = new Processor(uploadedFileDirectory, flag);
+                return processor.process(false);
+            }
+            else {
+                String uploadedFileDirectory = FILE_DIRECTORY + paramFile.getOriginalFilename();
+                Processor processor = new Processor(uploadedFileDirectory, flag);
+                return processor.process(true);
+            }
         }
         catch(IOException e) {
             System.out.println(e.getMessage());
         }
-        return "Something went wrong.";
-    }//end uploadDataset
-
-    @GetMapping("/process-dataset")
-    private DatasetEndpoint processDataset() {
-        try {
-            return Processor.process(UPLOADED_FILE_DIRECTORY);
-        }
-        catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-
         return null;
-    }//end processDataset
+    }//end uploadDataset
 
     @PostMapping("/process-custom")
     private ArrayList<CustomGraphLink> processCustom(@RequestBody CustomGraph customGraph) {
         com.jafar.wiring.custom.Processor processor = new com.jafar.wiring.custom.Processor(new ArrayList<>(Arrays.asList(customGraph.getLinks())), new ArrayList<>(Arrays.asList(customGraph.getNodes())));
         return processor.getLINKS();
     }//end processCustom
-
-
-
 }//end class
